address CR comments and update tests as needed
