+++ src/MSAL.PCL/Internal/Requests/SilentRequest.cs
@@ -28,6 +28,7 @@
 using System;
 using System.Threading.Tasks;
 using Microsoft.Identity.Client.Internal.OAuth2;
+using Microsoft.Identity.Client.Internal.Cache;
 
 namespace Microsoft.Identity.Client.Internal.Requests
 {
@@ -51,15 +52,50 @@ namespace Microsoft.Identity.Client.Internal.Requests
             throw new System.NotImplementedException();
         }
 
-        protected override Task SendTokenRequestAsync()
+        protected override async Task SendTokenRequestAsync()
         {
-            if (Exception == null)
+            if (!this.LoadFromCache)
             {
-                PlatformPlugin.Logger.Verbose(this.CallState, "No token matching arguments found in the cache");
-                throw new MsalSilentTokenAcquisitionException();
+                throw new MsalSilentTokenAcquisitionException(new Exception("Token cache is set to null"));
             }
 
-            throw new MsalSilentTokenAcquisitionException(Exception);
+            //look for access token first because force refresh is not set
+            if (!ForceRefresh)
+            {
+                AccessTokenItem
+                     = TokenCache.FindAccessToken(AuthenticationRequestParameters);
+            }
+
+            if (AccessTokenItem == null)
+            {
+                RefreshTokenCacheItem refreshTokenItem =
+                    TokenCache.FindRefreshToken(AuthenticationRequestParameters);
+
+                if (refreshTokenItem == null)
+                {
+                    PlatformPlugin.Logger.Verbose(this.CallState, "No token matching arguments found in the cache");
+                    throw new MsalSilentTokenAcquisitionException(
+                        new Exception("No token matching arguments found in the cache"));
+                }
+
+                await this.RefreshAccessTokenAsync(refreshTokenItem).ConfigureAwait(false);
+            }
+        }
+
+
+
+        internal async Task RefreshAccessTokenAsync(RefreshTokenCacheItem item)
+        {
+                PlatformPlugin.Logger.Verbose(this.CallState, "Refreshing access token...");
+                
+                    await this.SendTokenRequestByRefreshTokenAsync(item.RefreshToken).ConfigureAwait(false);
+
+                    if (Response.IdToken == null)
+                    {
+                        // If Id token is not returned by token endpoint when refresh token is redeemed, 
+                        // we should copy tenant and user information from the cached token.
+                        Response.IdToken = item.RawIdToken;
+                    }
         }
     }
 }
